+++ SimpleBank/Bank.cs
@@ -11,7 +11,6 @@ namespace SimpleBank
     /// </summary>
     public static class Bank
     {
-        //SDASD
         #region Public methods
         /// <summary>
         /// Process the loan request
@@ -21,11 +20,11 @@ namespace SimpleBank
         /// <param name="amount">The amount</param>
         /// <param name="duration">The duration</param>
         /// <returns>The interest rate</returns>
-        public static double ProcessLoanRequest(string ssn, int creditScore, double amount, int duration)
+        public static decimal ProcessLoanRequest(string ssn, int creditScore, decimal amount, int duration)
         {
             Random rnd = new Random();
 
-            return (((double)rnd.Next(10, 5999)) / 100);
+            return (((decimal)rnd.Next(10, 5999)) / 100);
         }
 
         #endregion
